feat: update pill list component
